Refactor containers to separate files
