+++ package.json
@@ -116,6 +116,7 @@
     "@types/react-dom": "0.14.18",
     "@types/react-measure": "0.4.5",
     "@types/react-redux": "4.4.38",
+    "@types/reselect": "^2.2.0",
     "@types/sinon": "1.16.32",
     "autoprefixer": "6.4.0",
     "classnames": "2.2.5",
